backend/DjangoRestApiMongoDB/requirements.txt
django
zappa
djangorestframework 
djangorestframework-simplejwt 